fixing missed int to long with zend_parse_parameters

+++ ext/ftp/php_ftp.c
@@ -17,7 +17,7 @@
    +----------------------------------------------------------------------+
  */
 
-/* $Id: php_ftp.c,v 1.74.2.5 2003/03/07 13:42:13 ddhill Exp $ */
+/* $Id: php_ftp.c,v 1.74.2.6 2003/03/10 15:57:07 ddhill Exp $ */
 
 #ifdef HAVE_CONFIG_H
 #include "config.h"
@@ -897,7 +897,7 @@ PHP_FUNCTION(ftp_nb_put)
 	ftptype_t	xtype;
 	char		*remote, *local;
 	int		remote_len, local_len, ret;
-	int		mode, startpos=0;
+	long		mode, startpos=0;
 	php_stream * instream;
 
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "rssl|l", &z_ftp, &remote, &remote_len, &local, &local_len, &mode, &startpos) == FAILURE) {
